extract components from TimePicker
